Split `DataGridContainer`
