+++ packages/admin/src/components/bindingFacade/collections/DataGrid/base/DataGridContainer/index.ts
@@ -1,0 +1,2 @@
+export * from './DataGridContainer'
+export * from './Types'
